+++ app/projects/forms.py
@@ -5,6 +5,7 @@ import json
 import io
 import csv
 from django.db.models import Q
+from django.utils.html import format_html
 from openpyxl import load_workbook
 import numpy as np
 
@@ -68,7 +69,6 @@ def gettext_variables(some_string, lang="de"):
 
 
 def add_help_text_icon(field, param_name, RTD_link=True):
-
     if field.help_text is not None:
         help_text = field.help_text + ". " + _("Click on the icon for more help") + "."
         field.help_text = None
@@ -769,9 +769,9 @@ class AssetCreateForm(OpenPlanModelForm):
                 "Electrical efficiency with no heat extraction"
             )
 
-            self.fields["efficiency"].help_text = (
-                "This is the custom help text for chp efficiency"
-            )
+            self.fields[
+                "efficiency"
+            ].help_text = "This is the custom help text for chp efficiency"
             self.add_help_text_icon("efficiency", RTD_link=True)
             self.fields["efficiency_multiple"] = DualNumberField(
                 default=1, min=0, max=1, param_name="efficiency_multiple"
@@ -786,9 +786,9 @@ class AssetCreateForm(OpenPlanModelForm):
             self.fields["efficiency"].label = _("Efficiency gas to electricity")
 
             # TODO
-            self.fields["efficiency"].help_text = (
-                "This is the custom help text for chp efficiency"
-            )
+            self.fields[
+                "efficiency"
+            ].help_text = "This is the custom help text for chp efficiency"
             self.add_help_text_icon("efficiency", RTD_link=True)
 
             self.fields["efficiency_multiple"].widget = forms.NumberInput(
@@ -812,9 +812,9 @@ class AssetCreateForm(OpenPlanModelForm):
                 }
             )
             self.fields["efficiency_multiple"].label = _("Heat loss")
-            self.fields["efficiency_multiple"].help_text = (
-                "Ratio of energy converted to heat"
-            )
+            self.fields[
+                "efficiency_multiple"
+            ].help_text = "Ratio of energy converted to heat"
             self.add_help_text_icon("efficiency_multiple", RTD_link=True)
 
         if "dso" in self.asset_type_name:
@@ -834,7 +834,7 @@ class AssetCreateForm(OpenPlanModelForm):
         """
         for field in self.fields:
             if field == "renewable_asset" and self.asset_type_name in RENEWABLE_ASSETS:
-                self.fields[field].initial = True
+                self.initial[field] = True
             self.fields[field].widget.attrs.update({f"df-{field}": ""})
             if field == "input_timeseries":
                 self.fields[field].required = self.is_input_timeseries_empty()
@@ -854,6 +854,8 @@ class AssetCreateForm(OpenPlanModelForm):
                 self.fields[field].label = self.fields[field].label.replace(
                     ":unit:", self.asset_type.unit
                 )
+
+            self.fields[field].label = format_html(self.fields[field].label)
 
         """ ----------------------------------------------------- """
 
@@ -1130,15 +1132,15 @@ class StorageForm(AssetCreateForm):
         asset_type_name = kwargs.pop("asset_type", None)
         super(StorageForm, self).__init__(*args, asset_type="capacity", **kwargs)
         self.fields["dispatchable"].widget = forms.HiddenInput()
-        self.fields["dispatchable"].initial = True
+        self.initial["dispatchable"] = True
 
         if asset_type_name != "hess":
             self.fields["fixed_thermal_losses_relative"].widget = forms.HiddenInput()
-            self.fields["fixed_thermal_losses_relative"].initial = 0
+            self.initial["fixed_thermal_losses_relative"] = 0
             self.fields["fixed_thermal_losses_absolute"].widget = forms.HiddenInput()
-            self.fields["fixed_thermal_losses_absolute"].initial = 0
+            self.initial["fixed_thermal_losses_absolute"] = 0
             self.fields["thermal_loss_rate"].widget = forms.HiddenInput()
-            self.fields["thermal_loss_rate"].initial = 0
+            self.initial["thermal_loss_rate"] = 0
         else:
             field_name = "fixed_thermal_losses_relative"
             help_text = self.fields[field_name].help_text