+++ app/projects/urls.py
@@ -143,6 +143,26 @@ urlpatterns = [
         name="scenario_duplicate",
     ),
     path("scenario/export/<int:proj_id>", scenario_export, name="scenario_export"),
+    path(
+        "scenario/export/datapackage/<int:scen_id>",
+        scenario_export_as_datapackage,
+        name="scenario_export_as_datapackage",
+    ),
+    path(
+        "scenario/export/datapackage/<int:scen_id>/number/<int:n_timestamps>",
+        scenario_export_as_datapackage,
+        name="scenario_export_as_datapackage",
+    ),
+    path(
+        "project/export/datapackage/<int:proj_id>",
+        project_export_as_datapackage,
+        name="project_export_as_datapackage",
+    ),
+    path(
+        "project/export/datapackage/<int:proj_id>/number/<int:n_timestamps>",
+        project_export_as_datapackage,
+        name="project_export_as_datapackage",
+    ),
     path("scenario/upload/<int:proj_id>", scenario_upload, name="scenario_upload"),
     # path('scenario/upload/<int:proj_id>', LoadScenarioFromFileView.as_view(), name='scenario_upload'),
     # Timeseries Model
@@ -155,7 +175,7 @@ urlpatterns = [
         name="get_constant_timeseries_id",
     ),
     re_path(
-        "get/constant/timeseries/id/(?P<ts_length>\d+)/value/(?P<value>\d+(\.\d+)?)/$",
+        r"get/constant/timeseries/id/(?P<ts_length>\d+)/value/(?P<value>\d+(\.\d+)?)/$",
         get_constant_timeseries_id,
         name="get_constant_timeseries_id",
     ),