+++ app/projects/views.py
@@ -1,4 +1,9 @@
 # from bootstrap_modal_forms.generic import BSModalCreateView
+import tempfile
+from pathlib import Path
+import zipfile
+
+
 from django.contrib.auth.decorators import login_required
 import datetime
 from django.http import (
@@ -111,7 +116,6 @@ def landing_default(request, version=1):
 @login_required
 @require_http_methods(["POST"])
 def scenario_upload(request, proj_id):
-
     # read the scenario file to a dict
     scenario_data = request.FILES["file"].read()
     scenario_data = json.loads(scenario_data)
@@ -149,7 +153,7 @@ def scenario_upload(request, proj_id):
         for i, scen in enumerate(scenario_data):
             if new_scenario_name != "":
                 if n_scenarios > 1:
-                    scen["name"] = f"{new_scenario_name}_{i+1}"
+                    scen["name"] = f"{new_scenario_name}_{i + 1}"
                 else:
                     scen["name"] = new_scenario_name
 
@@ -264,7 +268,6 @@ def project_revoke_access(request, proj_id=None):
 @json_view
 @require_http_methods(["POST"])
 def ajax_project_viewers_form(request):
-
     if request.headers.get("x-requested-with") == "XMLHttpRequest":
         proj_id = int(request.POST.get("proj_id"))
         project = get_object_or_404(Project, id=proj_id)
@@ -399,7 +402,6 @@ def project_update(request, proj_id):
                                     )
 
         if project_form.is_valid() and economic_data_form.is_valid():
-
             logger.info(f"Updating project with economic data...")
 
             project_form.save()
@@ -422,7 +424,6 @@ def project_update(request, proj_id):
 @login_required
 @require_http_methods(["GET", "POST"])
 def project_export(request, proj_id):
-
     project = get_object_or_404(Project, id=proj_id)
 
     if project.user != request.user:
@@ -448,7 +449,6 @@ def project_export(request, proj_id):
 @login_required
 @require_http_methods(["POST"])
 def project_upload(request):
-
     # read the project file to a dict
     project_data = request.FILES["file"].read()
     project_data = json.loads(project_data)
@@ -492,7 +492,6 @@ def project_from_usecase(request, usecase_id=None):
 
 @require_http_methods(["GET"])
 def usecase_export(request, usecase_id):
-
     usecase = get_object_or_404(UseCase, id=usecase_id)
 
     response = HttpResponse(
@@ -607,7 +606,6 @@ def project_duplicate(request, proj_id):
 # region Usecase
 @require_http_methods(["GET"])
 def usecase_search(request, usecase_id=None, scen_id=None):
-
     usecase_list = UseCase.objects.all()
 
     return render(
@@ -728,7 +726,6 @@ def scenario_search(request, proj_id, show_comments=0):
 def scenario_select_project(request, step_id=0, max_step=1):
     user_projects = fetch_user_projects(request.user)
     if request.method == "GET":
-
         if user_projects.exists():
             form = ScenarioSelectProjectForm(project_queryset=user_projects)
             answer = render(
@@ -762,7 +759,6 @@ def scenario_select_project(request, step_id=0, max_step=1):
 @login_required
 @require_http_methods(["GET", "POST"])
 def scenario_create_parameters(request, proj_id, scen_id=None, step_id=1, max_step=2):
-
     project = get_object_or_404(Project, pk=proj_id)
     # all projects which the user is able to select (the one the user created)
     user_projects = fetch_user_projects(request.user)
@@ -818,7 +814,6 @@ def scenario_create_parameters(request, proj_id, scen_id=None, step_id=1, max_st
         )
 
     elif request.method == "POST":
-
         form = ScenarioCreateForm(request.POST, project_queryset=user_projects)
 
         if form.is_valid():
@@ -835,8 +830,7 @@ def scenario_create_parameters(request, proj_id, scen_id=None, step_id=1, max_st
                 ).exists()
                 is True
             ):
-
-                qs_sim = Simulation.objects.filter(scenario=scenario)
+                qs_sim = Simulation.objects.filter(scenario__id=scenario.id)
                 # update the parameter values which are different from existing values
                 for name, value in form.cleaned_data.items():
                     # only update fields if they were changed
@@ -893,7 +887,6 @@ def scenario_create_parameters(request, proj_id, scen_id=None, step_id=1, max_st
 @login_required
 @require_http_methods(["GET", "POST"])
 def scenario_create_topology(request, proj_id, scen_id, step_id=2, max_step=3):
-
     components = {
         "providers": {
             "dso": _("Electricity DSO"),
@@ -976,7 +969,6 @@ def scenario_create_topology(request, proj_id, scen_id, step_id=2, max_step=3):
             # node_obj.assign_asset_to_proper_group(node_to_db_mapping_dict)
         return JsonResponse({"success": True}, status=200)
     else:
-
         scenario = get_object_or_404(Scenario, pk=scen_id)
 
         # if a simulation object linked to this scenario exists, all steps have been already fullfilled
@@ -1012,7 +1004,6 @@ def scenario_create_topology(request, proj_id, scen_id, step_id=2, max_step=3):
 @login_required
 @require_http_methods(["GET", "POST"])
 def scenario_create_constraints(request, proj_id, scen_id, step_id=3, max_step=4):
-
     constraints_labels = {
         "minimal_degree_of_autonomy": _("Minimal degree of autonomy"),
         "minimal_renewable_factor": _("Minimal share of renewables"),
@@ -1054,7 +1045,6 @@ def scenario_create_constraints(request, proj_id, scen_id, step_id=3, max_step=4
     qs_sim = Simulation.objects.filter(scenario=scenario)
 
     if request.method == "GET":
-
         # if a simulation object linked to this scenario exists, all steps have been already fullfilled
         if qs_sim.exists():
             max_step = 5
@@ -1231,7 +1221,6 @@ def scenario_review(request, proj_id, scen_id, step_id=4, max_step=MAX_STEP):
 @login_required
 @require_http_methods(["GET"])
 def back_to_scenario_review(request, proj_id):
-
     selected_scenario = get_selected_scenarios_in_cache(request, proj_id)
 
     if len(selected_scenario) >= 1:
@@ -1357,6 +1346,73 @@ def scenario_export(request, proj_id):
             json.dumps(scenario_data), content_type="application/json"
         )
         response["Content-Disposition"] = "attachment; filename=scenario.json"
+    return response
+
+
+@login_required
+@require_http_methods(["GET"])
+def scenario_export_as_datapackage(request, scen_id, n_timestamps=None):
+    scenario = get_object_or_404(Scenario, id=int(scen_id))
+
+    if scenario.project.user != request.user:
+        raise PermissionDenied
+
+    with tempfile.TemporaryDirectory() as temp_dir:
+        destination_path = Path(temp_dir)
+        # write the content of the scenario into a temp directory
+        scenario_folder = scenario.to_datapackage(destination_path, number=n_timestamps)
+
+        # Place the temp directory into a zip folder
+        zip_buffer = io.BytesIO()
+        with zipfile.ZipFile(zip_buffer, "w") as zip_file:
+            for file_path in destination_path.rglob("*"):  # Recursively walk all files
+                if file_path.is_file():
+                    # Relative path inside ZIP
+                    arcname = file_path.relative_to(destination_path)
+                    with open(file_path, "rb") as f:
+                        zip_file.writestr(str(arcname), f.read())
+
+        # Let the user download the zip file
+        zip_buffer.seek(0)
+        response = HttpResponse(zip_buffer.getvalue(), content_type="application/zip")
+        response["Content-Disposition"] = (
+            f'attachment; filename="datapackage_scenario_{scen_id}.zip"'
+        )
+
+    return response
+
+
+@login_required
+@require_http_methods(["GET"])
+def project_export_as_datapackage(request, proj_id, n_timestamps=None):
+    project = get_object_or_404(Project, id=int(proj_id))
+
+    if project.user != request.user:
+        raise PermissionDenied
+
+    with tempfile.TemporaryDirectory() as temp_dir:
+        destination_path = Path(temp_dir)
+
+        for scenario in project.scenario_set.all():
+            scenario.to_datapackage(destination_path, number=n_timestamps)
+
+        # Place the temp directory into a zip folder
+        zip_buffer = io.BytesIO()
+        with zipfile.ZipFile(zip_buffer, "w") as zip_file:
+            for file_path in destination_path.rglob("*"):  # Recursively walk all files
+                if file_path.is_file():
+                    # Relative path inside ZIP
+                    arcname = file_path.relative_to(destination_path)
+                    with open(file_path, "rb") as f:
+                        zip_file.writestr(str(arcname), f.read())
+
+        # Let the user download the zip file
+        zip_buffer.seek(0)
+        response = HttpResponse(zip_buffer.getvalue(), content_type="application/zip")
+        response["Content-Disposition"] = (
+            f'attachment; filename="datapackage_project_{proj_id}.zip"'
+        )
+
     return response
 
 
@@ -1677,7 +1733,6 @@ def get_asset_create_form(request, scen_id=0, asset_type_name="", asset_uuid=Non
             )
         return render(request, "asset/storage_asset_create_form.html", {"form": form})
     else:  # all other assets
-
         if asset_uuid:
             existing_asset = get_object_or_404(Asset, unique_id=asset_uuid)
             form = AssetCreateForm(
@@ -1761,7 +1816,6 @@ def asset_connection_ports_number(request, asset_type_name=None):
 @login_required
 @require_http_methods(["GET"])
 def asset_connection_ports_info(request, asset_type_name=None):
-
     energy_carrier = request.GET.get("energy_carrier", "Electricity").title()
     if asset_type_name is not None:
         if asset_type_name == "bus":
@@ -1868,7 +1922,6 @@ def view_mvs_data_input(request, scen_id=0, testing=False):
     scenario = Scenario.objects.get(id=scen_id)
 
     if scenario.project.user != request.user:
-
         logger.warning(
             f"Unauthorized user tried to access scenario with db id = {scen_id}."
         )
@@ -1877,7 +1930,6 @@ def view_mvs_data_input(request, scen_id=0, testing=False):
     try:
         data_clean = format_scenario_for_mvs(scenario, testing)
     except Exception as e:
-
         logger.error(
             f"Scenario Serialization ERROR! User: {scenario.project.user.username}. Scenario Id: {scenario.id}. Thrown Exception: {traceback.format_exc()}."
         )
@@ -1942,7 +1994,6 @@ def request_mvs_simulation(request, scen_id=0):
             content_type="application/json",
         )
     else:
-
         # delete existing simulation
         Simulation.objects.filter(scenario_id=scen_id).delete()