+++ app/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 ## [Unreleased]
 
+## [v2.1.1] – 2026-04-15
+### Fixed
+- Fix datapackage export if scenario name contains slashes [[#455](https://github.com/open-plan-tool/gui/pull/455)]
+
+## [v2.1.0] – 2026-04-14
+### Added
+- Enable download of datapackage as zip [[#422](https://github.com/open-plan-tool/gui/pull/422)] [[#447](https://github.com/open-plan-tool/gui/pull/447)]
+
+### Changed
+- Update Django version and dependencies [[#432](https://github.com/open-plan-tool/gui/pull/432)]
+
+### Fixed
+- Fix initial assignment in storage forms [[#435](https://github.com/open-plan-tool/gui/pull/435)]
+
 ## [v2.0.1] – 2026-01-27
 ### Changed
 - Updated favicon and project text to all say OpenPlan and use the most current logo.